chore: refactor sample composers - breakup class + renaming
Not updating any of the code - just moving things around
