+++ src/main/java/com/google/api/generator/gapic/composer/ClientLibraryPackageInfoComposer.java
@@ -21,7 +21,7 @@ import com.google.api.generator.engine.ast.JavaDocComment;
 import com.google.api.generator.engine.ast.PackageInfoDefinition;
 import com.google.api.generator.engine.ast.TypeNode;
 import com.google.api.generator.engine.ast.VaporReference;
-import com.google.api.generator.gapic.composer.samplecode.ServiceClientSampleCodeComposer;
+import com.google.api.generator.gapic.composer.samplecode.ServiceClientHeaderSampleComposer;
 import com.google.api.generator.gapic.composer.utils.ClassNames;
 import com.google.api.generator.gapic.model.GapicContext;
 import com.google.api.generator.gapic.model.GapicPackageInfo;
@@ -120,7 +120,7 @@ public class ClientLibraryPackageInfoComposer {
                   .setName(ClassNames.getServiceClientClassName(service))
                   .build());
       String packageInfoSampleCode =
-          ServiceClientSampleCodeComposer.composeClassHeaderMethodSampleCode(
+          ServiceClientHeaderSampleComposer.composeClassHeaderMethodSampleCode(
               service, clientType, context.resourceNames(), context.messages());
       javaDocCommentBuilder.addSampleCode(packageInfoSampleCode);
     }
