+++ src/main/java/com/google/api/generator/gapic/composer/common/AbstractServiceClientClassComposer.java
@@ -55,7 +55,9 @@ import com.google.api.generator.engine.ast.ValueExpr;
 import com.google.api.generator.engine.ast.Variable;
 import com.google.api.generator.engine.ast.VariableExpr;
 import com.google.api.generator.gapic.composer.comment.ServiceClientCommentComposer;
-import com.google.api.generator.gapic.composer.samplecode.ServiceClientSampleCodeComposer;
+import com.google.api.generator.gapic.composer.samplecode.ServiceClientCallableSampleComposer;
+import com.google.api.generator.gapic.composer.samplecode.ServiceClientHeaderSampleComposer;
+import com.google.api.generator.gapic.composer.samplecode.ServiceClientMethodSampleComposer;
 import com.google.api.generator.gapic.composer.store.TypeStore;
 import com.google.api.generator.gapic.composer.utils.ClassNames;
 import com.google.api.generator.gapic.composer.utils.PackageChecker;
@@ -189,13 +191,13 @@ public abstract class AbstractServiceClientClassComposer implements ClassCompose
     TypeNode clientType = typeStore.get(ClassNames.getServiceClientClassName(service));
     TypeNode settingsType = typeStore.get(ClassNames.getServiceSettingsClassName(service));
     String classMethodSampleCode =
-        ServiceClientSampleCodeComposer.composeClassHeaderMethodSampleCode(
+        ServiceClientHeaderSampleComposer.composeClassHeaderMethodSampleCode(
             service, clientType, resourceNames, messageTypes);
     String credentialsSampleCode =
-        ServiceClientSampleCodeComposer.composeClassHeaderCredentialsSampleCode(
+        ServiceClientHeaderSampleComposer.composeClassHeaderCredentialsSampleCode(
             clientType, settingsType);
     String endpointSampleCode =
-        ServiceClientSampleCodeComposer.composeClassHeaderEndpointSampleCode(
+        ServiceClientHeaderSampleComposer.composeClassHeaderEndpointSampleCode(
             clientType, settingsType);
     return ServiceClientCommentComposer.createClassHeaderComments(
         service, classMethodSampleCode, credentialsSampleCode, endpointSampleCode);
@@ -697,7 +699,7 @@ public abstract class AbstractServiceClientClassComposer implements ClassCompose
 
       Optional<String> methodSampleCode =
           Optional.of(
-              ServiceClientSampleCodeComposer.composeRpcMethodHeaderSampleCode(
+              ServiceClientMethodSampleComposer.composeRpcMethodHeaderSampleCode(
                   method, typeStore.get(clientName), signature, resourceNames, messageTypes));
       MethodDefinition.Builder methodVariantBuilder =
           MethodDefinition.builder()
@@ -777,7 +779,7 @@ public abstract class AbstractServiceClientClassComposer implements ClassCompose
 
     Optional<String> defaultMethodSampleCode =
         Optional.of(
-            ServiceClientSampleCodeComposer.composeRpcDefaultMethodHeaderSampleCode(
+            ServiceClientMethodSampleComposer.composeRpcDefaultMethodHeaderSampleCode(
                 method, typeStore.get(clientName), resourceNames, messageTypes));
 
     MethodInvocationExpr callableMethodExpr =
@@ -900,7 +902,7 @@ public abstract class AbstractServiceClientClassComposer implements ClassCompose
     if (callableMethodKind.equals(CallableMethodKind.LRO)) {
       sampleCodeOpt =
           Optional.of(
-              ServiceClientSampleCodeComposer.composeLroCallableMethodHeaderSampleCode(
+              ServiceClientCallableSampleComposer.composeLroCallableMethodHeaderSampleCode(
                   method,
                   typeStore.get(ClassNames.getServiceClientClassName(service)),
                   resourceNames,
@@ -908,7 +910,7 @@ public abstract class AbstractServiceClientClassComposer implements ClassCompose
     } else if (callableMethodKind.equals(CallableMethodKind.PAGED)) {
       sampleCodeOpt =
           Optional.of(
-              ServiceClientSampleCodeComposer.composePagedCallableMethodHeaderSampleCode(
+              ServiceClientCallableSampleComposer.composePagedCallableMethodHeaderSampleCode(
                   method,
                   typeStore.get(ClassNames.getServiceClientClassName(service)),
                   resourceNames,
@@ -917,7 +919,7 @@ public abstract class AbstractServiceClientClassComposer implements ClassCompose
       if (method.stream().equals(Stream.NONE)) {
         sampleCodeOpt =
             Optional.of(
-                ServiceClientSampleCodeComposer.composeRegularCallableMethodHeaderSampleCode(
+                ServiceClientCallableSampleComposer.composeRegularCallableMethodHeaderSampleCode(
                     method,
                     typeStore.get(ClassNames.getServiceClientClassName(service)),
                     resourceNames,
@@ -925,7 +927,7 @@ public abstract class AbstractServiceClientClassComposer implements ClassCompose
       } else {
         sampleCodeOpt =
             Optional.of(
-                ServiceClientSampleCodeComposer.composeStreamCallableMethodHeaderSampleCode(
+                ServiceClientCallableSampleComposer.composeStreamCallableMethodHeaderSampleCode(
                     method,
                     typeStore.get(ClassNames.getServiceClientClassName(service)),
                     resourceNames,
