+++ src/main/java/com/google/api/generator/gapic/composer/samplecode/SampleCodeWriter.java
@@ -38,7 +38,7 @@ public final class SampleCodeWriter {
     for (Statement statement : statements) {
       statement.accept(visitor);
     }
-    String formattedSampleCode = SampleCodeBodyJavaFormatter.format(visitor.write());
+    String formattedSampleCode = SampleBodyJavaFormatter.format(visitor.write());
     // Escape character "@" in the markdown code block <pre>{@code...} tags.
     return formattedSampleCode.replaceAll("@", "{@literal @}");
   }
