+++ src/test/java/com/google/api/generator/gapic/composer/samplecode/SampleCodeJavaFormatterTest.java
@@ -25,7 +25,7 @@ public class SampleCodeJavaFormatterTest {
   @Test
   public void validFormatSampleCode_tryCatchStatement() {
     String samplecode = LineFormatter.lines("try(boolean condition = false){", "int x = 3;", "}");
-    String result = SampleCodeBodyJavaFormatter.format(samplecode);
+    String result = SampleBodyJavaFormatter.format(samplecode);
     String expected =
         LineFormatter.lines("try (boolean condition = false) {\n", "  int x = 3;\n", "}");
     assertEquals(expected, result);
@@ -36,7 +36,7 @@ public class SampleCodeJavaFormatterTest {
     String sampleCode =
         "SubscriptionAdminSettings subscriptionAdminSettings = "
             + "SubscriptionAdminSettings.newBuilder().setEndpoint(myEndpoint).build();";
-    String result = SampleCodeBodyJavaFormatter.format(sampleCode);
+    String result = SampleBodyJavaFormatter.format(sampleCode);
     String expected =
         LineFormatter.lines(
             "SubscriptionAdminSettings subscriptionAdminSettings =\n",
@@ -50,7 +50,7 @@ public class SampleCodeJavaFormatterTest {
         "echoSettingsBuilder.echoSettings().setRetrySettings("
             + "echoSettingsBuilder.echoSettings().getRetrySettings().toBuilder()"
             + ".setTotalTimeout(Duration.ofSeconds(30)).build());";
-    String result = SampleCodeBodyJavaFormatter.format(sampleCode);
+    String result = SampleBodyJavaFormatter.format(sampleCode);
     String expected =
         LineFormatter.lines(
             "echoSettingsBuilder\n",
@@ -68,9 +68,9 @@ public class SampleCodeJavaFormatterTest {
   @Test
   public void invalidFormatSampleCode_nonStatement() {
     assertThrows(
-        SampleCodeBodyJavaFormatter.FormatException.class,
+        SampleBodyJavaFormatter.FormatException.class,
         () -> {
-          SampleCodeBodyJavaFormatter.format("abc");
+          SampleBodyJavaFormatter.format("abc");
         });
   }
 }
